backend/requirements.txt
chromadb==1.2.1
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
sentence-transformers==2.2.2
torch==2.2.0
transformers==5.0.0  